requirements.txt
asgiref==3.7.2
asttokens==2.4.1
autopep8==2.0.4
backports.zoneinfo==0.2.1
black==23.11.0
click==8.1.7
colorama==0.4.6
Django==4.2.3
django-debug-toolbar==4.2.0
django-filter==23.3
django-silk==5.0.4
django-simple-history==3.4.0
executing==2.0.1
flake8==6.1.0
gprof2dot==2022.7.29
icecream==2.1.3
mccabe==0.7.0
more-itertools==10.1.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.17.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
transliterate==1.10.2
typing_extensions==4.7.1
tzdata==2023.3
